Move cross-platform tile machinery out of apple/
`Sources/GeoDrawer/apple/` is supposed to hold only code that's
Apple-only by nature. Several pure-Swift pieces had crept in: the
`Sampling` enum, the `BaseMap`/`BaseMapImage` value+class shells,
the `TileCache`/`TileCacheKey` pair, and the entire `tilesNeeded`
/ `prefetchTiles` / `resolvedZoom` machinery. Linux's CI caught the
inconsistency the hard way — `TiledBaseMap.swift` (top-level)
referenced `BaseMap.Sampling` which lived inside the CG-gated
`apple/GeoDrawer+BaseMap.swift`, so Linux builds failed with
"cannot find type 'BaseMap' in scope".

Reorg:

- New `Sources/GeoDrawer/Sampling.swift` — `GeoDrawer.Sampling`
  (`nearest`/`bilinear`). Was `BaseMap.Sampling`; both
  single-image and tiled base maps use it now.
- New `Sources/GeoDrawer/BaseMap.swift` — `BaseMap` struct +
  `BaseMapImage` class (with a pure-Swift `init?(width:height:pixels:)`
  for callers like swift-png on Linux). The `Hashable` conformance
  lives here too.
- `apple/BaseMap+CoreGraphics.swift` (renamed from
  `apple/GeoDrawer+BaseMap.swift`) keeps only the CGImage decoder
  on `BaseMapImage` and the `CGImage`/`UIImage`/`NSImage`
  initialisers on `BaseMap`.
- New `Sources/GeoDrawer/TileCache.swift` — `TileCache` class +
  `TileCacheKey`. Pure Swift, NSLock-serialised.
- New `Sources/GeoDrawer/GeoDrawer+TilePrefetch.swift` —
  `resolvedZoom`, `tilesNeeded`, `prefetchTiles`, and the
  `TileFetchOutcome` private struct. The previous version called
  `baseMapCache.invalidateTiled` directly; now there's an
  `invalidateRenderedTiledRaster(matching:)` shim that's a no-op
  on Linux and clears the apple-side raster cache when CG is
  present.
- `apple/GeoDrawer+TiledBaseMap.swift` keeps only the CG-bound
  renderer (`renderedTiledBaseMap`, `renderTiledBaseMap`,
  `TiledRasterCacheKey`, `TiledRasterContext`).
- `GeoDrawer.swift`: `tileCache` and `pixelDensity` properties
  moved out of the `#if canImport(CoreGraphics)` block — both are
  used by the pure-Swift prefetch path. `baseMapCache` stays
  gated (it caches `CGImage`).

`TiledBaseMap.sampling` and `BaseMap.sampling` now both use
`GeoDrawer.Sampling`. No behaviour change on Apple; Linux now
compiles. All 86 tests still pass; both Cassini targets build clean.
